+++ src/yb/docdb/doc_kv_util.h
@@ -47,7 +47,7 @@ CHECKED_STATUS DecodeHybridTimeFromEndOfKey(const rocksdb::Slice &key, DocHybrid
 // encoded DocHybridTime is ValueType::kHybridTime.
 CHECKED_STATUS CheckHybridTimeSizeAndValueType(
     const rocksdb::Slice& key,
-    int* ht_byte_size_dest);
+    size_t* ht_byte_size_dest);
 
 // Consumes hybrid time from the given slice, decreasing the slice size by the hybrid time size.
 // Hybrid time is stored in a "key-appropriate" format (bits inverted for reverse sorting).
