+++ src/yb/docdb/doc_pgsql_scanspec.cc
@@ -172,7 +172,7 @@ DocPgsqlScanSpec::DocPgsqlScanSpec(
     InitRangeOptions(*condition);
 
     // Range options are only valid if all range columns are set (i.e. have one or more options).
-    for (int i = 0; i < schema_.num_range_key_columns(); i++) {
+    for (size_t i = 0; i < schema_.num_range_key_columns(); i++) {
       if ((*range_options_)[i].empty()) {
         range_options_ = nullptr;
         break;
